@@ -10,24 +10,31 @@
         "production": "mix --production"
     },
     "devDependencies": {
-        "axios": "^0.21",
-        "bootstrap": "^4.0.0",
         "jquery": "^3.2",
-        "laravel-mix": "^6.0.6",
+        "laravel-mix": "^6.0.19",
         "lodash": "^4.17.19",
         "popper.js": "^1.12",
         "postcss": "^8.1.14",
         "resolve-url-loader": "^2.3.1",
-        "sass": "^1.20.1",
-        "sass-loader": "^8.0.0",
-        "vue": "^2.5.17",
-        "vue-loader": "^15.9.5",
-        "vue-template-compiler": "^2.6.10"
+        "sass": "^1.34.0",
+        "sass-loader": "^11.1.1",
+        "vue-loader": "^15.9.7",
+        "vue-template-compiler": "^2.6.12"
     },
     "dependencies": {
+        "@fortawesome/fontawesome-free": "^5.13.0",
+        "@popperjs/core": "^2.9.2",
         "animate.css": "^4.1.1",
+        "axios": "^0.19.2",
+        "bootstrap": "^4.6.0",
+        "bootstrap-vue": "^2.21.2",
         "cross-env": "^7.0.3",
-        "vue-axios": "^3.2.4",
-        "vue-router": "^3.5.1"
+        "sweetalert2": "^9.10.12",
+        "toastr": "^2.1.4",
+        "vue": "^2.6.12",
+        "vue-axios": "^2.1.4",
+        "vue-router": "^3.5.1",
+        "vuelidate": "^0.7.5",
+        "vuex": "^3.3.0"
     }
 }
