@@ -15,8 +15,7 @@ class GroupController extends Controller
     public function index()
     {
         $groups = Group::all();
-        
-        return view ('groups.index', compact('groups'));
+        return $groups;
     }
 
     /**
@@ -26,7 +25,7 @@ class GroupController extends Controller
      */
     public function create()
     {
-        return view('groups.create');
+        //
     }
 
     /**
@@ -37,11 +36,11 @@ class GroupController extends Controller
      */
     public function store(Request $request)
     {
-        $input = $request->all();
+        $data = $request->all();
+        $group = new Group($data);
+        $group->save();
 
-        Group::create($input);
-
-        return redirect()->route('group.index');
+        return $group;
     }
 
     /**
@@ -50,9 +49,9 @@ class GroupController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Group $group)
     {
-        //
+        return $group;
     }
 
     /**
@@ -63,7 +62,7 @@ class GroupController extends Controller
      */
     public function edit(Group $group)
     {
-        return view('groups.edit', compact('group'));
+        //
     }
 
     /**
@@ -75,10 +74,8 @@ class GroupController extends Controller
      */
     public function update(Request $request, Group $group)
     {
-       
         $group->update($request->all());
-
-        return redirect()->route('group.index');
+        return $group;
     }
 
     /**
@@ -90,9 +87,8 @@ class GroupController extends Controller
     public function destroy(Group $group)
     {
         $group->delete();
-
         return response()->json([
             'message' => 'Grupo Eliminado'
-          ]); 
+        ]);
     }
 }
