@@ -13,21 +13,23 @@
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Poppins:300,400,500,600,700" />
     <!--end::Fonts-->
     <!--begin::Page Vendors Styles(used by this page)-->
-    <link href="{{ asset('metronic/assets/plugins/custom/fullcalendar/fullcalendar.bundle.css?v=7.0.4') }}"
-        rel="stylesheet" type="text/css" />
+    <!-- <link href="{{ asset('metronic/assets/plugins/custom/fullcalendar/fullcalendar.bundle.css?v=7.0.4') }}"
+        rel="stylesheet" type="text/css" /> -->
     <!--end::Page Vendors Styles-->
     <!--begin::Global Theme Styles(used by all pages)-->
-    <link href="{{ asset('metronic/assets/plugins/global/plugins.bundle.css?v=7.0.4') }}" rel="stylesheet"
-        type="text/css" />
-    <link href="{{ asset('metronic/assets/plugins/custom/prismjs/prismjs.bundle.css?v=7.0.4') }}" rel="stylesheet"
-        type="text/css" />
-    <link href="{{ asset('metronic/assets/css/style.bundle.css?v=7.0.4') }}" rel="stylesheet" type="text/css" />
-    <script src="https://code.jquery.com/jquery-2.2.4.min.js"
-        integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44=" crossorigin="anonymous"></script>
+    <!-- <link href="{{ asset('metronic/assets/plugins/global/plugins.bundle.css?v=7.0.4') }}" rel="stylesheet"
+        type="text/css" /> -->
+    <!-- <link href="{{ asset('metronic/assets/plugins/custom/prismjs/prismjs.bundle.css?v=7.0.4') }}" rel="stylesheet"
+        type="text/css" /> -->
+    <!-- <link href="{{ asset('metronic/assets/css/style.bundle.css?v=7.0.4') }}" rel="stylesheet" type="text/css" /> -->
+    <!-- <script src="https://code.jquery.com/jquery-2.2.4.min.js"
+        integrity="sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44=" crossorigin="anonymous"></script> -->
     <!--end::Global Theme Styles-->
     <!--begin::Layout Themes(used by all pages)-->
     <!--end::Layout Themes-->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/latest/toastr.min.css"></script>
+    <!-- Scripts -->
+    <link rel="stylesheet" href="{{ asset('css/app.css') }}">
+    <script src="{{ asset('js/app.js') }}" defer></script>
 </head>
 <!--end::Head-->
 <!--begin::Body-->
@@ -35,93 +37,31 @@
 <body id="kt_body" style="background-image: url({{ asset('metronic/assets/media/bg/bg-10.jpg') }}"
     class="quick-panel-right demo-panel-right offcanvas-right header-fixed subheader-enabled page-loading">
 
-    @include('layouts.metronic.header_mobile')
+    <div id="app">
 
-    <div class="d-flex flex-column flex-root">
-        <!--begin::Page-->
-        <div class="d-flex flex-row flex-column-fluid page">
-            <!--begin::Wrapper-->
-            <div class="d-flex flex-column flex-row-fluid wrapper" id="kt_wrapper">
-                <!--begin::Header-->
-                @include('layouts.metronic.header')
-                <!--end::Header-->
-                <!--begin::Content-->
-                <div class="content d-flex flex-column flex-column-fluid" id="kt_content">
-                    <!--begin::Subheader-->
-                    @include('layouts.metronic.subheader')
-                    <!--end::Subheader-->
-                    <!--begin::Entry-->
-                    <div class="d-flex flex-column-fluid">
-                        <!--begin::Container-->
-                        <div class="container">
-                            @yield('content')
-                        </div>
-                        <!--end::Container-->
-                    </div>
-                    <!--end::Entry-->
-                </div>
-                <!--end::Content-->
-                <!--begin::Footer-->
-                @include('layouts.metronic.footer')
-                <!--end::Footer-->
-            </div>
-            <!--end::Wrapper-->
+        @include('layouts.metronic.header_mobile')
+
+        @yield('content')
+
+        <!--begin::Scrolltop-->
+        <div id="kt_scrolltop" class="scrolltop">
+            <span class="svg-icon">
+                <!--begin::Svg Icon | path:assets/media/svg/icons/Navigation/Up-2.svg-->
+                <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px"
+                    height="24px" viewBox="0 0 24 24" version="1.1">
+                    <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
+                        <polygon points="0 0 24 0 24 24 0 24" />
+                        <rect fill="#000000" opacity="0.3" x="11" y="10" width="2" height="10" rx="1" />
+                        <path
+                            d="M6.70710678,12.7071068 C6.31658249,13.0976311 5.68341751,13.0976311 5.29289322,12.7071068 C4.90236893,12.3165825 4.90236893,11.6834175 5.29289322,11.2928932 L11.2928932,5.29289322 C11.6714722,4.91431428 12.2810586,4.90106866 12.6757246,5.26284586 L18.6757246,10.7628459 C19.0828436,11.1360383 19.1103465,11.7686056 18.7371541,12.1757246 C18.3639617,12.5828436 17.7313944,12.6103465 17.3242754,12.2371541 L12.0300757,7.38413782 L6.70710678,12.7071068 Z"
+                            fill="#000000" fill-rule="nonzero" />
+                    </g>
+                </svg>
+                <!--end::Svg Icon-->
+            </span>
         </div>
-        <!--end::Page-->
+        <!--end::Scrolltop-->
     </div>
-    <!--end::Main-->
-    <!--begin::Quick Panel-->
-
-    <!--end::Quick Panel-->
-    <!--begin::Scrolltop-->
-    <div id="kt_scrolltop" class="scrolltop">
-        <span class="svg-icon">
-            <!--begin::Svg Icon | path:assets/media/svg/icons/Navigation/Up-2.svg-->
-            <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="24px"
-                height="24px" viewBox="0 0 24 24" version="1.1">
-                <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-                    <polygon points="0 0 24 0 24 24 0 24" />
-                    <rect fill="#000000" opacity="0.3" x="11" y="10" width="2" height="10" rx="1" />
-                    <path
-                        d="M6.70710678,12.7071068 C6.31658249,13.0976311 5.68341751,13.0976311 5.29289322,12.7071068 C4.90236893,12.3165825 4.90236893,11.6834175 5.29289322,11.2928932 L11.2928932,5.29289322 C11.6714722,4.91431428 12.2810586,4.90106866 12.6757246,5.26284586 L18.6757246,10.7628459 C19.0828436,11.1360383 19.1103465,11.7686056 18.7371541,12.1757246 C18.3639617,12.5828436 17.7313944,12.6103465 17.3242754,12.2371541 L12.0300757,7.38413782 L6.70710678,12.7071068 Z"
-                        fill="#000000" fill-rule="nonzero" />
-                </g>
-            </svg>
-            <!--end::Svg Icon-->
-        </span>
-    </div>
-    <!--end::Scrolltop-->
-    <!--begin::Sticky Toolbar-->
-    <ul class="sticky-toolbar nav flex-column pl-2 pr-2 pt-3 pb-3 mt-4">
-        <!--begin::Item-->
-        <li class="nav-item mb-2" id="kt_demo_panel_toggle" data-toggle="tooltip" title="Check out more demos"
-            data-placement="right">
-            <a class="btn btn-sm btn-icon btn-bg-light btn-icon-success btn-hover-success" href="#">
-                <i class="flaticon2-drop"></i>
-            </a>
-        </li>
-        <!--end::Item-->
-        <!--begin::Item-->
-        <li class="nav-item mb-2" data-toggle="tooltip" title="Layout Builder" data-placement="left">
-            <a class="btn btn-sm btn-icon btn-bg-light btn-icon-primary btn-hover-primary"
-                href="https://keenthemes.com/metronic/preview/demo2/builder.html" target="_blank">
-                <i class="flaticon2-gear"></i>
-            </a>
-        </li>
-        <!--end::Item-->
-        <!--begin::Item-->
-        <li class="nav-item mb-2" data-toggle="tooltip" title="Documentation" data-placement="left">
-            <a class="btn btn-sm btn-icon btn-bg-light btn-icon-warning btn-hover-warning"
-                href="https://keenthemes.com/metronic/?page=docs" target="_blank">
-                <i class="flaticon2-telegram-logo"></i>
-            </a>
-        </li>
-        <!--end::Item-->
-    </ul>
-    <!--end::Sticky Toolbar-->
-    <!--begin::Demo Panel-->
-
-    <!--end::Demo Panel-->
     <script>
         var HOST_URL = "https://keenthemes.com/metronic/tools/preview";
 
@@ -190,15 +130,15 @@
     </script>
     <!--end::Global Config-->
     <!--begin::Global Theme Bundle(used by all pages)-->
-    <script src="{{ asset('metronic/assets/plugins/global/plugins.bundle.js?v=7.0.4') }}"></script>
-    {{-- <script src="{{ asset('metronic/assets/plugins/custom/prismjs/prismjs.bundle.js?v=7.0.4') }}"></script> --}}
-    <script src="{{ asset('metronic/assets/js/scripts.bundle.js?v=7.0.4') }}"></script>
+    <!-- <script src="{{ asset('metronic/assets/plugins/global/plugins.bundle.js?v=7.0.4') }}"></script> -->
+    <!-- {{-- <script src="{{ asset('metronic/assets/plugins/custom/prismjs/prismjs.bundle.js?v=7.0.4') }}"></script> --}} -->
+    <!-- <script src="{{ asset('metronic/assets/js/scripts.bundle.js?v=7.0.4') }}"></script> -->
     <!--end::Global Theme Bundle-->
     <!--begin::Page Vendors(used by this page)-->
-    <script src="{{ asset('metronic/assets/plugins/custom/fullcalendar/fullcalendar.bundle.js?v=7.0.4') }}"></script>
+    <!-- <script src="{{ asset('metronic/assets/plugins/custom/fullcalendar/fullcalendar.bundle.js?v=7.0.4') }}"></script> -->
     <!--end::Page Vendors-->
     <!--begin::Page Scripts(used by this page)-->
-    <script src="{{ asset('metronic/assets/js/pages/widgets.js?v=7.0.4') }}"></script>
+    <!-- <script src="{{ asset('metronic/assets/js/pages/widgets.js?v=7.0.4') }}"></script> -->
 
     <!--end::Page Scripts-->
 </body>
